requirements.txt
langchain 
streamlit  
langchain_experimental 
python_dotenv 
langchain_community 
tabulate 
langchain_openai